Avoid warnings about signed/unsigned comparison in wxGrid code
Explicitly cast wxNumberFormatter::GetDecimalSeparator() to int before
comparing it with int keycode.

Master contains a better fix for this problem (see #23103), but this one
is simpler and works also in non-Unicode build.

See #23101.

+++ src/generic/grideditors.cpp
@@ -1049,7 +1049,7 @@ void wxGridCellFloatEditor::StartingKey(wxKeyEvent& event)
     int keycode = event.GetKeyCode();
 
     if ( wxIsdigit(keycode) || keycode == '+' || keycode == '-'
-         || keycode == wxNumberFormatter::GetDecimalSeparator() )
+         || keycode == static_cast<int>(wxNumberFormatter::GetDecimalSeparator()) )
     {
         wxGridCellTextEditor::StartingKey(event);
 
@@ -1186,7 +1186,7 @@ bool wxGridCellFloatEditor::IsAcceptedKey(wxKeyEvent& event)
             // accept digits, 'e' as in '1e+6', also '-', '+', and '.'
             if ( wxIsdigit(keycode) ||
                     tolower(keycode) == 'e' ||
-                        keycode == wxNumberFormatter::GetDecimalSeparator() ||
+                        keycode == static_cast<int>(wxNumberFormatter::GetDecimalSeparator()) ||
                             keycode == '+' ||
                                 keycode == '-' )
             {
